Frontend mistakes corrected
Now the case is decrypted directly from server, and the server is responsible for license transfer. As user need to login and get the key of a case from server, its useless to decrypt on the frontend. Though does allow the use of decryption in frontend by private key. But not required anymore.