CIP Hash + Documents Encryption
CIPs are hashed as other document, and encrypted with all of the available documents with their respective type.
Aka passport will be encrypted as password type, CIP will be encrypted as CIP type. While this will be the only document which has been hashed. So the only way to merge cases, is if the other case also has same CIP at the moment of insertion.