Cases Licenses now created with condition
If the email creator of the case is present in DRM database, the license will also be generated for it. If not just for the user configured manager from config.
Also now the registrations of new account is connected with godata server, users cannot register with email/password other than the one of godata.